feat: add lead time for change metric
The assumption here is that the first commit designates work start.
This is highly unlikely and data is destroyed or falsified when either
squashing or rebasing the branch.

This works for a first iteration though but requires external data we cannot
hope to get e.g. from pull request or task management system.